+++ src/controls/listItemPicker/IListItemPickerState.ts
@@ -1,4 +1,4 @@
-export interface IFieldPickerListDataState {
+export interface IListItemPickerState {
   noresultsFoundText: string;
   showError: boolean;
   errorMessage: string;